use vertex llm as base class for embeddings
